Feature/phenopp yield (#63)
* better support for models with orphan metabolites

previously, if a model had orphan metabolites then those did not get any
constraints added by to_solver_based_model. Those metabolites could then
not have id changed etc as no key in the constraints dictionary. Fix
that by adding zero-constraints to all metabolits that are not part of
any reaction.

* report mass/c yield in phenotypic phase plane

allow interpretation of the feeding source-to-product yield in mass (g
product / g source) and (mol C product / mol C product) when examining
the phenotypic phase plane plot showing a selected variable reaction as
defined flux and another reaction that is being optimized.

* fix corner cases for phpp yield

return 0 when there is no medium defined or if the medium reports
multiple carbon sources

* re-organize

n_carbon better methods of metabolite and reaction. avoid redundant []
in sum([..])

* correct doc style

use google style doc as rest of cameo

* fix doc strings, use get

* wip

* report mass/c yield in phenotypic phase plane

allow interpretation of the feeding source-to-product yield in mass (g
product / g source) and (mol C product / mol C product) when examining
the phenotypic phase plane plot showing a selected variable reaction as
defined flux and another reaction that is being optimized.

* fix corner cases for phpp yield

return 0 when there is no medium defined or if the medium reports
multiple carbon sources

* re-organize

n_carbon better methods of metabolite and reaction. avoid redundant []
in sum([..])

* correct doc style

use google style doc as rest of cameo

* fix doc strings, use get

* Fix: gene.knock_out method (#69)

* fix: Gene.knock_out for complex gene-protein-reaction association rules

This was done by setting the _functional_ attribute of the gene to False and
 cheking if this has an effect on the gene-rule evaluation. The reaction is only
 knocked out when all it's gene rule evaluates False.

* update test

* fix tests / revert column names

* remove stray copy/paste

* debug test

* debug tests for carbon/mass yield

* fix: return nan where nan is appropriate

- make n_carbon property
- prettier axis labels

* fix: avoid math.nan

* fix: remove bad error catch

* fix: typo, reaction list should not return 0

* fix: exceptions

- glpk bug fixed so don't catch UndefinedSolution
- don't try/except shield more statments than necessary
- return nan instead of 0 if fail to define source reactions

* fix: skip pointless yield calculation

if model solve didn't work, don't try to estimate other yields either

* chore: make travis report to slack less frequently (#83)

* re-organize

n_carbon better methods of metabolite and reaction. avoid redundant []
in sum([..])

* fix doc strings, use get

* re-organize

n_carbon better methods of metabolite and reaction. avoid redundant []
in sum([..])

* fix doc strings, use get

* Fix: gene.knock_out method (#69)

* fix: Gene.knock_out for complex gene-protein-reaction association rules

This was done by setting the _functional_ attribute of the gene to False and
 cheking if this has an effect on the gene-rule evaluation. The reaction is only
 knocked out when all it's gene rule evaluates False.

* fix: return nan where nan is appropriate

- make n_carbon property
- prettier axis labels

* fix: misc

- (trial for evaluation) update the c-source reactions at all points as
  these may change depending on the fixed variables
- reduce digits in unit-test reduce noise
- allow giving variables as reactions

* fix: merge mistake

* fix: change implementation, source as parameters

auto computing the source reaction proved messy and hard to interpret as
the source might change over the course of the production
envelope. Instead, be explicit and have user-defined source with a
default to the reaction with max carbon flux

* fix: pep8